+++ src/Normalization/EolNormalization.php
@@ -85,7 +85,7 @@ class EolNormalization implements NormalizationInterface, ConfigurableNormalizat
         return $file;
     }
 
-    private function getConfig(string $key, mixed $default): mixed
+    private function getConfig(string $key, mixed $default = null): mixed
     {
         if (isset($this->overrides[$key])) {
             return $this->overrides[$key];