+++ src/Normalization/FinalEolNormalization.php
@@ -99,7 +99,7 @@ class FinalEolNormalization implements NormalizationInterface, ConfigurableNorma
         return $file;
     }
 
-    private function getConfig(string $key, mixed $default): mixed
+    private function getConfig(string $key, mixed $default = null): mixed
     {
         if (isset($this->overrides[$key])) {
             return $this->overrides[$key];