+++ CHANGELOG.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- ...
+- Encoding can be manually entered when cannot be confidently retrieved
+  automatically.
 
 ## [0.0.2] - 2023-06-19